Move most of scylla_cql::value to core
The only thing that remains in deser_cql_value because we have no
intention of moving it.
