+++ scylla-cql-core/src/lib.rs
@@ -3,3 +3,8 @@
 //! This crate contains CQL data types and traits that form part of the stable
 //! public API of the `scylla` driver crate. It is re-exported by both `scylla`
 //! and `scylla-cql`, ensuring type identity is preserved across versions.
+
+pub(crate) mod pretty;
+pub(crate) mod utils;
+
+pub mod value;
