+++ scylla-cql-core/src/utils/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod safe_format;
