+++ scylla-cql-core/src/utils/safe_format.rs
