+++ scylla-cql/Cargo.toml
@@ -33,15 +33,15 @@ secrecy-08 = ["dep:secrecy-08"]
 # Enables support for CQL ser/de of Secrecy type from secrecy 0.10 crate.
 secrecy-10 = ["dep:secrecy-10"]
 # Enables support for CQL ser/de of time/date types from time 0.3 crate.
-time-03 = ["dep:time-03"]
+time-03 = ["dep:time-03", "scylla-cql-core/time-03"]
 # Enables support for CQL ser/de of time/date types from chrono 0.4 crate.
-chrono-04 = []
+chrono-04 = ["scylla-cql-core/chrono-04"]
 # Enables support for CQL ser/de of arbitrary length integer types from num-bigint 0.3 crate.
-num-bigint-03 = ["dep:num-bigint-03"]
+num-bigint-03 = ["dep:num-bigint-03", "scylla-cql-core/num-bigint-03"]
 # Enables support for CQL ser/de of arbitrary length integer types from num-bigint 0.4 crate.
-num-bigint-04 = ["dep:num-bigint-04"]
+num-bigint-04 = ["dep:num-bigint-04", "scylla-cql-core/num-bigint-04"]
 # Enables support for CQL ser/de of arbitrary precision decimal types from bigdecimal 0.4 crate.
-bigdecimal-04 = ["dep:bigdecimal-04"]
+bigdecimal-04 = ["dep:bigdecimal-04", "scylla-cql-core/bigdecimal-04"]
 # Enables support for CQL ser/de of all supported external types.
 full-serialization = [
     "chrono-04",
