+++ scylla-cql/src/utils/mod.rs
@@ -2,4 +2,3 @@
 // TODO(2.0): Deconstruct this module by extracting submodules to the top level.
 
 pub mod parse;
-pub(crate) mod safe_format;
